genesis-agent/backupmanager-agent/app/agent.py
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import datetime
import os
from zoneinfo import ZoneInfo

import google.auth
from google.adk.agents import Agent

_, project_id = google.auth.default()
os.environ.setdefault("GOOGLE_CLOUD_PROJECT", project_id)
os.environ.setdefault("GOOGLE_CLOUD_LOCATION", "global")
os.environ.setdefault("GOOGLE_GENAI_USE_VERTEXAI", "True")



def web_search(query: str) -> str:
    """Search the web for current information.
    
    Args:
        query: The search query
        
    Returns:
        Search results as a string
    """
    # TODO: Implement real web search (Serper API, Google Search API, etc.)
    return f"Searched for: {query}. (Web search implementation pending)"


def generate_code(language: str, description: str) -> str:
    """Generate code in any language.
    
    Args:
        language: Programming language (python, javascript, etc.)
        description: What the code should do
        
    Returns:
        Generated code
    """
    return f"# Generated {language} code for: {description}\n# (Code generation via LLM)"


def analyze_code(code: str) -> str:
    """Analyze code for issues and improvements.
    
    Args:
        code: The code to analyze
        
    Returns:
        Analysis results
    """
    return f"Code analysis: (Analysis via LLM)"


def execute_python(code: str) -> str:
    """Execute Python code safely.
    
    Args:
        code: Python code to execute
        
    Returns:
        Execution output
    """
    # TODO: Implement sandboxed execution
    return "Python execution: (Pending sandbox implementation)"


def call_api(url: str, method: str = "GET", **kwargs) -> str:
    """Call an external API.
    
    Args:
        url: API endpoint
        method: HTTP method
        
    Returns:
        API response
    """
    return f"API call to {url}: (Implementation pending)"


def transcribe_audio(audio_data: str) -> str:
    """Transcribe audio to text.
    
    Args:
        audio_data: Base64 encoded audio
        
    Returns:
        Transcribed text
    """
    return "Audio transcription: (Using Gemini audio capabilities)"


def analyze_image(image_data: str) -> str:
    """Analyze an image.
    
    Args:
        image_data: Base64 encoded image
        
    Returns:
        Image analysis
    """
    return "Image analysis: (Using Gemini vision)"


def generate_image(prompt: str) -> str:
    """Generate an image from text.
    
    Args:
        prompt: Image description
        
    Returns:
        Generated image data
    """
    return "Image generation: (Using Imagen or similar)"


def save_note(content: str, tags: str = "") -> str:
    """Save a note.
    
    Args:
        content: Note content
        tags: Optional tags
        
    Returns:
        Confirmation
    """
    return f"Note saved with tags: {tags}"


def search_notes(query: str) -> str:
    """Search saved notes.
    
    Args:
        query: Search query
        
    Returns:
        Matching notes
    """
    return f"Searching notes for: {query}"


def send_email(to: str, subject: str, body: str) -> str:
    """Send an email.
    
    Args:
        to: Recipient
        subject: Email subject
        body: Email body
        
    Returns:
        Confirmation
    """
    return f"Email sent to {to}"


def manage_calendar(action: str, **kwargs) -> str:
    """Manage calendar events.
    
    Args:
        action: create, update, delete, list
        
    Returns:
        Action result
    """
    return f"Calendar action: {action}"


def upload_file(filename: str, data: str) -> str:
    """Upload a file.
    
    Args:
        filename: File name
        data: File data
        
    Returns:
        Upload confirmation
    """
    return f"File uploaded: {filename}"


def organize_files(pattern: str, action: str) -> str:
    """Organize files.
    
    Args:
        pattern: File pattern
        action: Organization action
        
    Returns:
        Organization result
    """
    return f"Files organized: {pattern}"


def search_files(query: str) -> str:
    """Search for files.
    
    Args:
        query: Search query
        
    Returns:
        Matching files
    """
    return f"Searching files: {query}"


def create_workflow(name: str, steps: str) -> str:
    """Create an automation workflow.
    
    Args:
        name: Workflow name
        steps: Workflow steps
        
    Returns:
        Workflow created confirmation
    """
    return f"Workflow '{name}' created"


def scrape_web(url: str) -> str:
    """Scrape data from a webpage.
    
    Args:
        url: Website URL
        
    Returns:
        Extracted data
    """
    return f"Scraped data from: {url}"


def summarize_text(text: str) -> str:
    """Summarize text.
    
    Args:
        text: Text to summarize
        
    Returns:
        Summary
    """
    return "Text summary: (Using LLM)"


def analyze_sentiment(text: str) -> str:
    """Analyze sentiment and tone.
    
    Args:
        text: Text to analyze
        
    Returns:
        Sentiment analysis (positive/negative/neutral, tone, emotion)
    """
    return "Sentiment analysis: (Using LLM for emotion detection)"


def generate_text(prompt: str) -> str:
    """Generate text.
    
    Args:
        prompt: Generation prompt
        
    Returns:
        Generated text
    """
    return "Generated text: (Using LLM)"


def analyze_video(video_data: str) -> str:
    """Analyze video content.
    
    Args:
        video_data: Video data
        
    Returns:
        Video analysis
    """
    return "Video analysis: (Using Gemini video)"


def text_to_speech(text: str) -> str:
    """Convert text to speech.
    
    Args:
        text: Text to convert
        
    Returns:
        Audio data
    """
    return "Text-to-speech: (Implementation pending)"


def index_data(data: str, metadata: str) -> str:
    """Index data for search.
    
    Args:
        data: Data to index
        metadata: Metadata
        
    Returns:
        Index confirmation
    """
    return "Data indexed"


def monitor_system() -> str:
    """Check system health.
    
    Returns:
        System status
    """
    return "System status: Healthy"


def debug_agent(agent_name: str) -> str:
    """Debug an agent.
    
    Args:
        agent_name: Agent to debug
        
    Returns:
        Debug info
    """
    return f"Debugging {agent_name}"


root_agent = Agent(
    name="BackupManager",
    model="gemini-2.5-flash",
    instruction="""You are BackupManager - a backup and recovery specialist.
        
🎯 YOUR EXPERTISE:
- Backup strategies
- Disaster recovery
- Data archiving
- Snapshot management

🛠️ YOUR TOOLS:
- upload_file: Backup files
- call_api: Use cloud backup APIs
- organize_files: Organize backups

Ensure data is always safe and recoverable.""",
    tools=[upload_file, call_api, organize_files],
)
